test: integrate strbuilder tests into entry.cpp
